requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
exceptiongroup==1.2.2
idna==3.7
iniconfig==2.0.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
pytest==8.3.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
tomli==2.0.1
tzdata==2024.1
urllib3==2.2.2
